[worker] Refactor daemon_data to prepare v3
instanceStatus is replaced by instanceData because it may be constructed
from config, monitor, status
